CORE-15575: addressed code review comments
