+++ kotlin-samples/oracle-foreign-exchange/.ci/nightly/JenkinsfileSnykScan
