+++ kotlin-samples/oracle-foreign-exchange/.run/runConfigurations/DebugCorDapp.run.xml
