+++ kotlin-samples/oracle-foreign-exchange/.snyk
