+++ kotlin-samples/oracle-foreign-exchange/config/gradle-plugin-default-key.pem
