+++ kotlin-samples/oracle-foreign-exchange/config/log4j2.xml
