+++ kotlin-samples/oracle-foreign-exchange/config/r3-ca-key.pem
