+++ kotlin-samples/oracle-foreign-exchange/config/static-network-config.json
