+++ kotlin-samples/oracle-foreign-exchange/gradle.properties
