+++ kotlin-samples/oracle-foreign-exchange/settings.gradle
@@ -20,7 +20,7 @@ pluginManagement {
 }
 
 // Root project name, used in naming the project as a whole and used in naming objects built by the project.
-rootProject.name = 'corda5-foreign-exchange-kotlin'
+rootProject.name = 'corda5-oracle-foreign-exchange-kotlin'
 include ':workflows'
 include ':contracts'
