+++ kotlin-samples/oracle-foreign-exchange/workflows/src/main/kotlin/com/r3/developers/samples/fx/workflows/QuoteExchangeRateSubFlow.kt
@@ -6,8 +6,17 @@ import net.corda.v5.application.flows.InitiatingFlow
 import net.corda.v5.application.flows.ResponderFlow
 import net.corda.v5.application.flows.SubFlow
 import net.corda.v5.application.messaging.FlowSession
+import net.corda.v5.base.annotations.CordaSerializable
 import net.corda.v5.base.annotations.Suspendable
+import net.corda.v5.base.types.MemberX500Name
 import org.slf4j.LoggerFactory
+import java.math.BigDecimal
+
+@CordaSerializable
+data class QuoteFxRateRequest(val currencyPair: String, val fxServiceName: MemberX500Name)
+
+@CordaSerializable
+data class QuoteFxRateResponse(val conversionRate: BigDecimal)
 
 @InitiatingFlow(protocol = "quote-fx")
 class QuoteExchangeRateSubFlow(
